pyproject.toml
[project]
name = "person-validator"
version = "0.1.0"
description = "Web service and application to identify and enrich publicly accessible data about persons."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.2",
    "gunicorn>=25.1.0",
    "python-ulid>=3.1.0",
    "typing-extensions>=4.15.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "pytest-django>=4.12.0",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
DJANGO_SETTINGS_MODULE = "src.web.config.settings"
django_find_project = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
show_missing = true
fail_under = 80

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
